Fix includes (printf etc). Remove compiler warnings for wrong format specifier where we don't explicitly test it.
git-svn-id: http://root.cern.ch/svn/root/trunk@31785 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/test/t1054.cxx
@@ -16,8 +16,8 @@ typedef unsigned long long  ULong64_t;
 
 
 int main() {
-  printf("%d\n",sizeof(Long64_t));
-  printf("%d\n",sizeof(ULong64_t));
+   printf("%d\n",(int)sizeof(Long64_t));
+   printf("%d\n",(int)sizeof(ULong64_t));
   return 0;
 }
